# backend/static/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"TableDetail": {
				"properties": {
					"Name": {
						"type": "string"
					},
					"Description": {
						"type": "string"
					},
					"State": {
						"type": "string"
					}
				},
				"required": [
					"Name",
					"Description"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"keycloak": {
				"type": "apiKey",
				"name": "Authorization",
				"description": "Keycloak Bearer Token",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/Table": {
			"get": {
				"operationId": "GetTableAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"State": {
												"type": "string"
											},
											"Description": {
												"type": "string"
											},
											"Name": {
												"type": "string"
											},
											"Id": {
												"type": "string"
											}
										},
										"required": [
											"State",
											"Description",
											"Name",
											"Id"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateTable",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [
					{
						"keycloak": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TableDetail"
							}
						}
					}
				}
			}
		},
		"/Table/{id}": {
			"get": {
				"operationId": "GetTableById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"allOf": [
												{
													"properties": {
														"Order": {
															"items": {
																"properties": {
																	"TableId": {
																		"type": "string"
																	},
																	"TypeService": {
																		"type": "string"
																	},
																	"TotalPrice": {
																		"type": "number",
																		"format": "double"
																	},
																	"OrderDate": {
																		"type": "string"
																	},
																	"EmployeeId": {
																		"type": "string"
																	},
																	"CustomerId": {
																		"type": "string"
																	},
																	"State": {
																		"type": "string"
																	},
																	"Id": {
																		"type": "string"
																	}
																},
																"required": [
																	"TableId",
																	"TypeService",
																	"TotalPrice",
																	"OrderDate",
																	"EmployeeId",
																	"CustomerId",
																	"State",
																	"Id"
																],
																"type": "object"
															},
															"type": "array"
														}
													},
													"required": [
														"Order"
													],
													"type": "object"
												},
												{
													"properties": {
														"State": {
															"type": "string"
														},
														"Description": {
															"type": "string"
														},
														"Name": {
															"type": "string"
														},
														"Id": {
															"type": "string"
														}
													},
													"required": [
														"State",
														"Description",
														"Name",
														"Id"
													],
													"type": "object"
												}
											]
										},
										{
											"type": "string",
											"enum": [
												"ไม่พบข้อมูล"
											]
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateTable",
				"responses": {
					"200": {
						"description": "Update",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [
					{
						"keycloak": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TableDetail"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteTable",
				"responses": {
					"200": {
						"description": "Delete",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [
					{
						"keycloak": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}